utilities: split `maybe` to own file
remove circular dependency of utilities -> alloc -> concurrent_stack ->
utilities
