+++ pbbslib/concurrent_stack.h
@@ -37,7 +37,7 @@
 #include <cstdint>
 #include <iostream>
 
-#include "utilities.h"
+#include "maybe.h"
 
 template<typename T>
 class concurrent_stack {
